--- data/main/2024-moon-data.json
@@ -1,1 +1,0 @@
-{}
